feat/trends 🔥: refactored some functions away from trending and tokens component
