+++ components/TokenCard.tsx
@@ -1,4 +1,5 @@
 import { blurHashPlaceholder } from '@/constants/App'
+import { formatPriceChange, formatValue } from '@/libs/string.helpers'
 import { cn } from '@/libs/utils'
 import { BirdEyeTokenItem } from '@/types'
 import { Image } from 'expo-image'
@@ -39,54 +40,6 @@ export function TokenCard({
     }
   }
 
-  const formatBalance = (balance: number, symbol?: string) => {
-    if (balance >= 1000000000) {
-      return `${(balance / 1000000000).toFixed(2)}B ${symbol || ''}`
-    } else if (balance >= 1000000) {
-      return `${(balance / 1000000).toFixed(2)}M ${symbol || ''}`
-    } else if (balance >= 1000) {
-      return `${(balance / 1000).toFixed(2)}K ${symbol || ''}`
-    } else if (balance < 0.01 && balance > 0) {
-      // Show up to two significant digits after leading zeros, no exponential
-      const balanceStr = balance.toFixed(8) // up to 8 decimals for safety
-      const match = balanceStr.match(/^0\.0*(\d{1,2})/)
-      const digits = match
-        ? match[1]
-        : balanceStr.split('.')[1]?.slice(0, 2) || '00'
-      return `0.${'0'.repeat(balanceStr.split('.')[1]?.indexOf(digits) ?? 0)}${digits} ${symbol || ''}`
-    } else {
-      return `${balance.toFixed(balance >= 1 ? 2 : 4)} ${symbol || ''}`
-    }
-  }
-
-  const formatValue = (value?: number) => {
-    if (!value) return '$0.00'
-
-    if (value >= 1000000000) {
-      return `$${(value / 1000000000).toFixed(2)}B`
-    } else if (value >= 1000000) {
-      return `$${(value / 1000000).toFixed(2)}M`
-    } else if (value >= 1000) {
-      return `$${(value / 1000).toFixed(2)}K`
-    } else if (value < 0.01 && value > 0) {
-      // Show zeros and first two non-zero digits for small values, e.g. 0.00034 -> $0.00034
-      const valueStr = value.toFixed(8) // up to 8 decimals for safety
-      const match = valueStr.match(/^0\.0*(\d{1,2})/)
-      const digits = match
-        ? match[1]
-        : valueStr.split('.')[1]?.slice(0, 2) || '00'
-      return `$0.${'0'.repeat(valueStr.split('.')[1]?.indexOf(digits) ?? 0)}${digits}`
-    } else {
-      return `$${value.toFixed(2)}`
-    }
-  }
-
-  const formatPriceChange = (change?: number) => {
-    if (!change) return '+0.0%'
-    const sign = change >= 0 ? '+' : ''
-    return `${sign}${change.toFixed(2)}%`
-  }
-
   // const imageUri = token.logoURI || token.icon
   const imageUri = token.logoURI
   const showImage = imageUri && !imageError
@@ -123,7 +76,7 @@ export function TokenCard({
             <Text className='text-gray-400 text-sm' numberOfLines={1}>
               {/* {token.name || 'Unknown Token'} */}
               {mc && '$'}
-              {formatBalance(token.uiAmount || token.balance, token.symbol)}
+              {formatValue(token.uiAmount || token.balance)} {token.symbol}
             </Text>
             {/* {showBalance && (
               <Text className='text-gray-500 text-xs'>
